User controller and API updates
This commit includes updates which adds tests for the new backend user
controller and the API controller in turn then uses those functions to
do all of its operations.

This commit also includes slight changes to routes.js to align with the
new naming scheme.

The workout API and workout controller test also had a slight
modification as to close the server at the end of the tests.

I’ve tested and run on my machine and all tests are passing.